requirements.txt
bcrypt==4.3.0
mysql-connector-python==9.3.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
fastapi==0.109.2
python-multipart==0.0.6
uvicorn==0.34.0
reportlab==4.1.0
python-jose==3.3.0
databases==0.8.0
aiomysql==0.2.0
python-dotenv==1.0.0
email-validator==2.1.0
